+++ ChangeLog.md
@@ -4,6 +4,7 @@
 - NEW : Ajout de la class TechATM pour l'affichage de la page "A propos" *10/05/2022* 3.2.0
 
 ## 3.1
+- FIX: Compatibility V16 : newToken - *02/06/2022* - 3.1.5
 - FIX: save for wrong line - *2022-03-29* - 3.1.4
 - FIX: v14 compatibility quick extrafield - *2022-02-07* - 3.1.3
 - FIX: minor v15 compatibility issues - *2022-01-31* - 3.1.2